backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
google-generativeai
pydantic
python-dotenv
